+++ ModFixes/Fixes.cs
@@ -29,6 +29,8 @@ namespace Polyfill.ModFixes
             new S1MapiPrefabs(),
             new S1MapiInstancedTrees(),
             new OverTheCounterNetworkLib(),
+            new OverTheCounterDrifterPrefab(),
+            new BiggerTreesLodBias(),
         };
 
         private static MelonPreferences_Entry<string> _disabled;